+++ docs/annotated_dup.js
@@ -1,11 +1,14 @@
 var annotated_dup =
 [
     [ "termstyle", "namespacetermstyle.html", [
-      [ "OnExit", "classtermstyle_1_1_on_exit.html", null ],
+      [ "Col256", "structtermstyle_1_1_col256.html", "structtermstyle_1_1_col256" ],
+      [ "Config", "structtermstyle_1_1_config.html", "structtermstyle_1_1_config" ],
+      [ "OnExit", "classtermstyle_1_1_on_exit.html", "classtermstyle_1_1_on_exit" ],
       [ "PresetConfig", "structtermstyle_1_1_preset_config.html", null ],
       [ "StyledCout", "classtermstyle_1_1_styled_cout.html", null ],
-      [ "StyleString", "structtermstyle_1_1_style_string.html", null ]
+      [ "StyleString", "structtermstyle_1_1_style_string.html", "structtermstyle_1_1_style_string" ]
     ] ],
+    [ "BadColorID", "class_bad_color_i_d.html", null ],
     [ "Error", "class_error.html", null ],
     [ "PresetNameUsed", "class_preset_name_used.html", null ],
     [ "PresetNotFound", "class_preset_not_found.html", null ]