+++ docs/group___codes.js
@@ -1,0 +1,5 @@
+var group___codes =
+[
+    [ "termstyle::Codes", "group___codes.html#ga483e716a292eb17055b9d1632bc3e2f8", null ],
+    [ "termstyle::code2string", "group___codes.html#gaf81ff8310ccfdb964f85043689815e99", null ]
+];