+++ docs/group___error.js
@@ -2,5 +2,6 @@ var group___error =
 [
     [ "Error", "class_error.html", null ],
     [ "PresetNameUsed", "class_preset_name_used.html", null ],
-    [ "PresetNotFound", "class_preset_not_found.html", null ]
+    [ "PresetNotFound", "class_preset_not_found.html", null ],
+    [ "BadColorID", "class_bad_color_i_d.html", null ]
 ];