+++ docs/search/all_10.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['style_0',['style',['../namespacetermstyle.html#adc9b81309bb371263cf81f9c3a42eacf',1,'termstyle']]],
+  ['styledcout_1',['StyledCout',['../classtermstyle_1_1_styled_cout.html',1,'termstyle']]],
+  ['stylestring_2',['StyleString',['../structtermstyle_1_1_style_string.html',1,'termstyle']]]
+];