+++ docs/search/all_12.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['unreleased_20a_0',['&lt;a href=&quot;https://github.com/mrmagic2020/termstyle/compare/v1.0.0-pre.1...HEAD&quot; &gt;Unreleased&lt;/a&gt;',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md1',1,'']]],
+  ['usage_1',['Usage',['../index.html#autotoc_md16',1,'Basic usage'],['../index.html#autotoc_md8',1,'Usage']]],
+  ['using_20a_20preset_2',['Using a preset',['../index.html#autotoc_md13',1,'']]]
+];