+++ docs/search/all_13.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['v1_200_200_20pre_201_201_200_200_20pre_201_20a_202024_2004_2016_0',['&lt;a href=&quot;https://github.com/mrmagic2020/termstyle/releases/tag/v1.0.0-pre.1&quot; &gt;1.0.0-pre.1&lt;/a&gt; - 2024-04-16',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md5',1,'']]],
+  ['v1_200_200_20pre_201_20head_20unreleased_20a_1',['&lt;a href=&quot;https://github.com/mrmagic2020/termstyle/compare/v1.0.0-pre.1...HEAD&quot; &gt;Unreleased&lt;/a&gt;',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md1',1,'']]],
+  ['validatecolorid_2',['validateColorID',['../group__validate_color_i_d.html#ga3da712bc12b9a89917d79ec09c5fb72c',1,'termstyle']]]
+];