+++ docs/search/all_14 2.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['_7eonexit_0',['~OnExit',['../classtermstyle_1_1_on_exit.html#afb8dc15c160ceb03cf39a1145ccb3b6b',1,'termstyle::OnExit']]]
+];