+++ docs/search/all_2.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['error_0',['Error',['../class_error.html',1,'']]],
-  ['errors_1',['Errors',['../group___error.html',1,'']]]
+  ['2024_2004_2016_0',['&lt;a href=&quot;https://github.com/mrmagic2020/termstyle/releases/tag/v1.0.0-pre.1&quot; &gt;1.0.0-pre.1&lt;/a&gt; - 2024-04-16',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md5',1,'']]]
 ];