+++ docs/search/all_6.js
@@ -1,8 +1,7 @@
 var searchData=
 [
-  ['presetconfig_0',['PresetConfig',['../structtermstyle_1_1_preset_config.html',1,'termstyle']]],
-  ['presetnameused_1',['PresetNameUsed',['../class_preset_name_used.html',1,'']]],
-  ['presetnotfound_2',['PresetNotFound',['../class_preset_not_found.html',1,'']]],
-  ['presets_3',['presets',['../namespacetermstyle.html#a08f12dc9dd3354ab5175240f7f2368a9',1,'termstyle']]],
-  ['print_4',['print',['../namespacetermstyle.html#a5291f41cbd1f184e612f421343497a41',1,'termstyle']]]
+  ['error_0',['Error',['../class_error.html',1,'']]],
+  ['errors_1',['Errors',['../group___error.html',1,'']]],
+  ['example_2',['Example',['../index.html#autotoc_md10',1,'Example'],['../index.html#autotoc_md12',1,'Example'],['../index.html#autotoc_md14',1,'Example']]],
+  ['examples_3',['Examples',['../index.html#autotoc_md15',1,'']]]
 ];