+++ docs/search/all_8.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['termstyle_0',['termstyle',['../namespacetermstyle.html',1,'termstyle'],['../index.html',1,'termstyle']]],
-  ['termstyle_3a_3aconfig_1',['config',['../namespacetermstyle_1_1config.html',1,'termstyle']]]
+  ['github_20com_20mrmagic2020_20termstyle_20compare_20v1_200_200_20pre_201_20head_20unreleased_20a_0',['&lt;a href=&quot;https://github.com/mrmagic2020/termstyle/compare/v1.0.0-pre.1...HEAD&quot; &gt;Unreleased&lt;/a&gt;',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md1',1,'']]],
+  ['github_20com_20mrmagic2020_20termstyle_20releases_20tag_20v1_200_200_20pre_201_201_200_200_20pre_201_20a_202024_2004_2016_1',['&lt;a href=&quot;https://github.com/mrmagic2020/termstyle/releases/tag/v1.0.0-pre.1&quot; &gt;1.0.0-pre.1&lt;/a&gt; - 2024-04-16',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md5',1,'']]]
 ];