+++ docs/search/all_a.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['id_0',['ID',['../structtermstyle_1_1_col256.html#a992d1b18558171efe65fa9ab73789793',1,'termstyle::Col256']]],
+  ['input_1',['Fancy Input',['../index.html#autotoc_md20',1,'']]]
+];