+++ docs/search/all_d 2.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['onexit_0',['OnExit',['../classtermstyle_1_1_on_exit.html',1,'termstyle']]],
+  ['output_1',['Output',['../index.html#autotoc_md17',1,'Output'],['../index.html#autotoc_md19',1,'Output'],['../index.html#autotoc_md21',1,'Output']]]
+];