+++ docs/search/all_f 2.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['registering_20a_20preset_0',['Registering a preset',['../index.html#autotoc_md11',1,'']]],
+  ['releases_20tag_20v1_200_200_20pre_201_201_200_200_20pre_201_20a_202024_2004_2016_1',['&lt;a href=&quot;https://github.com/mrmagic2020/termstyle/releases/tag/v1.0.0-pre.1&quot; &gt;1.0.0-pre.1&lt;/a&gt; - 2024-04-16',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md5',1,'']]],
+  ['removed_2',['Removed',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md3',1,'']]]
+];