+++ docs/search/classes_3.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['styledcout_0',['StyledCout',['../classtermstyle_1_1_styled_cout.html',1,'termstyle']]],
-  ['stylestring_1',['StyleString',['../structtermstyle_1_1_style_string.html',1,'termstyle']]]
+  ['onexit_0',['OnExit',['../classtermstyle_1_1_on_exit.html',1,'termstyle']]]
 ];