+++ docs/search/files_0.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['termstyle_2ehpp_0',['termstyle.hpp',['../termstyle_8hpp.html',1,'']]]
+];