+++ docs/search/namespaces_0.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['termstyle_0',['termstyle',['../namespacetermstyle.html',1,'']]],
-  ['termstyle_3a_3aconfig_1',['config',['../namespacetermstyle_1_1config.html',1,'termstyle']]]
+  ['termstyle_0',['termstyle',['../namespacetermstyle.html',1,'']]]
 ];