+++ docs/structtermstyle_1_1_col256.js
@@ -1,0 +1,5 @@
+var structtermstyle_1_1_col256 =
+[
+    [ "ID", "structtermstyle_1_1_col256.html#a992d1b18558171efe65fa9ab73789793", null ],
+    [ "mode", "structtermstyle_1_1_col256.html#a80f1a6b440dbdf6a13a6248be9ac9f59", null ]
+];