+++ docs/structtermstyle_1_1_style_string 2.js
@@ -1,0 +1,8 @@
+var structtermstyle_1_1_style_string =
+[
+    [ "poststyle256", "structtermstyle_1_1_style_string.html#a20662057735761efb61bef569167ccbf", null ],
+    [ "poststyles", "structtermstyle_1_1_style_string.html#a8240d394862887395e80c4cc9e2173cc", null ],
+    [ "prestlye256", "structtermstyle_1_1_style_string.html#ab59899d7e8f651f5d92f1a45c9b49e28", null ],
+    [ "prestyles", "structtermstyle_1_1_style_string.html#ae48320a63a8c75dd7807f2d79642e0f2", null ],
+    [ "text", "structtermstyle_1_1_style_string.html#ab6cfdf27cb1c6a83c931a15fe26aa658", null ]
+];