+++ docs/topics.js
@@ -1,4 +1,6 @@
 var topics =
 [
-    [ "Errors", "group___error.html", "group___error" ]
+    [ "Errors", "group___error.html", "group___error" ],
+    [ "Col16", "group___codes.html", "group___codes" ],
+    [ "Col256", "group__validate_color_i_d.html", "group__validate_color_i_d" ]
 ];