+++ CHANGELOG.md
@@ -7,6 +7,26 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [1.0.0-pre.2] - 2024-04-17
+
+### Added
+
+- Support for Color256.
+
+- Configuration for automatic new lines after printing.
+
+### Removed
+
+- Removed namespace `termstyle::config`.
+
+### Fixed
+
+- Fix an issue where `poststyles` may be applied after a new line character.
+
+- STL include issues.
+
+- Improve code readability.
+
 ## [1.0.0-pre.1] - 2024-04-16
 
 - initial release
@@ -16,5 +36,6 @@ and this project adheres to [Semantic Versioning].
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/mrmagic2020/termstyle/compare/v1.0.0-pre.1...HEAD
-[1.0.0-pre.1]: https://github.com/mrmagic2020/termstyle/releases/tag/v1.0.0-pre.1
+[unreleased]: https://github.com/mrmagic2020/termstyle/compare/v1.0.0-pre.2...HEAD
+[1.0.0-pre.2]: https://github.com/mrmagic2020/termstyle/releases/v1.0.0-pre.2
+[1.0.0-pre.1]: https://github.com/mrmagic2020/termstyle/releases/v1.0.0-pre.1